test: adjust test code
