+++ .github/workflows/publish.yml
@@ -4,6 +4,9 @@ on:
   push:
     branches:
       - main
+  pull_request:
+    branches:
+      - main
 
 jobs:
   check-version:
@@ -19,9 +22,11 @@ jobs:
           deno-version: v2.x
       - run: deno fmt --check
       - run: deno lint
+      - run: deno task check
 
   publish:
     needs: check
+    if: github.event_name != 'pull_request'
     runs-on: ubuntu-latest
     permissions:
       contents: read
@@ -32,6 +37,7 @@ jobs:
 
   release:
     needs: check
+    if: github.event_name != 'pull_request'
     runs-on: ubuntu-latest
     permissions:
       contents: write
