+++ test/deno.json
@@ -4,6 +4,11 @@
     "noImplicitAny": false,
     "strictNullChecks": false
   },
+  "lint": {
+    "rules": {
+      "exclude": ["no-import-prefix"]
+    }
+  },
   "tasks": {
     "test": "deno test -A test.js",
     "update-snapshots": "deno test -A test.js -- --update",
