+++ test/test.js
@@ -1,116 +1,9 @@
-import { assertSnapshot } from "https://deno.land/std@0.185.0/testing/snapshot.ts";
-
-const run = async (cmd, cwd) => {
-  const command = new Deno.Command("sh", {
-    args: ["-c", cmd],
-    stdout: "piped",
-    stderr: "piped",
-    cwd: cwd,
-  });
-
-  const { code, stdout, stderr } = await command.output();
-
-  const removeAnsi = (input) => {
-    // deno-lint-ignore no-control-regex
-    const ansiEscapeSequences = /\u001b\[[0-9;]*[a-zA-Z]/g;
-    return input.replace(ansiEscapeSequences, "");
-  };
-
-  const removeVaryingOutput = (input) => {
-    return input
-      .replace(/obtained in \d+ms/g, "obtained in *ms")
-      .replace(/date: [^\n]+/g, "date: *")
-      .replace(/content-length: [^\n]+/g, "content-length: *");
-  };
-
-  let output = new TextDecoder().decode(stdout);
-  let outputError = new TextDecoder().decode(stderr);
-
-  output = removeVaryingOutput(removeAnsi(output));
-  outputError = removeAnsi(outputError);
-
-  return {
-    code,
-    output,
-    outputError,
-  };
-};
+import { createSdkTestRunner, createTestRunner } from "./test-utils.js";
 
 Deno.test("Given API", async (t) => {
-  const apiCommand = new Deno.Command("deno", {
-    args: "run -A test-api.js".split(" "),
-    stdout: "piped",
-    stderr: "piped",
-  });
-
-  const apiProcess = apiCommand.spawn();
-
-  do {
-    const outputReader = apiProcess.stdout.getReader();
-    const apiOutput = new TextDecoder().decode(
-      (await outputReader.read()).value,
-    );
-    outputReader.releaseLock();
-    if (apiOutput.indexOf("Listening on port 3000") != -1) {
-      break;
-    }
-
-    await new Promise((resolve) => setTimeout(resolve, 100));
-  } while (true);
-
-  const test = async (jsonrCommand, cwd) => {
-    let mainPath = "../main.js";
-    let absoluteCwd;
-    if (cwd) {
-      const projectRoot = Deno.cwd().replace("/test", "");
-      absoluteCwd = `${projectRoot}/${cwd}`;
-      if (cwd.includes("api1") || cwd.includes("api2")) {
-        mainPath = "../../../main.js";
-      }
-    }
-    await t.step(jsonrCommand + (cwd ? ` [cwd: ${cwd}]` : ""), async () => {
-      const { code, output, outputError } = await run(
-        jsonrCommand.replace("jsonr", `deno run -A ${mainPath}`),
-        absoluteCwd,
-      );
-
-      await assertSnapshot(t, { jsonrCommand, code, output, outputError });
-    });
-  };
-
-  const sdkTest = async (jsonrCommand) => {
-    await t.step(jsonrCommand, async () => {
-      const initResult = await run(jsonrCommand.replace("jsonr", "deno run -A ../main.js"));
-
-      const scriptResult = await run(
-        "deno run -A ../main.js run jsonr-script.js",
-      );
-
-      const normalizeScriptOutput = (output) => {
-        return output
-          .replace(/Elapsed: \d+ ms/g, "Elapsed: * ms")
-          .replace(/Header date: [^\n]+/g, "Header date: *");
-      };
-
-      try {
-        await assertSnapshot(t, {
-          jsonrCommand,
-          initCode: initResult.code,
-          initOutput: initResult.output,
-          initOutputError: initResult.outputError,
-          scriptCode: scriptResult.code,
-          scriptOutput: normalizeScriptOutput(scriptResult.output),
-          scriptOutputError: scriptResult.outputError,
-        });
-      } finally {
-        try {
-          await Deno.remove("jsonr-script.js");
-        } catch {
-          // Ignore if file doesn't exist
-        }
-      }
-    });
-  };
+  const apiProcess = await startTestApi();
+  const test = createTestRunner(t);
+  const sdkTest = createSdkTestRunner(t);
 
   await test("jsonr -e test get.http", "test/requests/api2");
   await test("jsonr -e test post.http", "test/requests/api2");
@@ -150,3 +43,28 @@ Deno.test("Given API", async (t) => {
   apiProcess.kill();
   await apiProcess.output();
 });
+
+const startTestApi = async () => {
+  const apiCommand = new Deno.Command("deno", {
+    args: "run -A test-api.js".split(" "),
+    stdout: "piped",
+    stderr: "piped",
+  });
+
+  const apiProcess = apiCommand.spawn();
+
+  do {
+    const outputReader = apiProcess.stdout.getReader();
+    const apiOutput = new TextDecoder().decode(
+      (await outputReader.read()).value,
+    );
+    outputReader.releaseLock();
+    if (apiOutput.indexOf("Listening on port 3000") != -1) {
+      break;
+    }
+
+    await new Promise((resolve) => setTimeout(resolve, 100));
+  } while (true);
+
+  return apiProcess;
+};
